+++ examples/minimal-consumer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # minimal-consumer
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @pwrdrvr/agent-client@0.5.0
+
 ## 0.0.3
 
 ### Patch Changes